// src/app/api/model/chargeData.ts
/**
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements. See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership. The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License. You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied. See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */

/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { PaymentTypeData } from './paymentTypeData';
import { CurrencyData } from './currencyData';
import { GLAccountData } from './gLAccountData';
import { TaxGroupData } from './taxGroupData';
import { EnumOptionData } from './enumOptionData';
import { ChargeFeeOnMonthDay } from './chargeFeeOnMonthDay';


export interface ChargeData { 
    accountMappingForChargeConfig?: string;
    active?: boolean;
    amount?: number;
    assetAccountOptions?: Array<GLAccountData>;
    chargeAppliesTo?: EnumOptionData;
    chargeAppliesToOptions?: Array<EnumOptionData>;
    chargeCalculationType?: EnumOptionData;
    chargeCalculationTypeOptions?: Array<EnumOptionData>;
    chargePaymentMode?: EnumOptionData;
    chargePaymetModeOptions?: Array<EnumOptionData>;
    chargeTimeType?: EnumOptionData;
    chargeTimeTypeOptions?: Array<EnumOptionData>;
    clientChargeCalculationTypeOptions?: Array<EnumOptionData>;
    clientChargeTimeTypeOptions?: Array<EnumOptionData>;
    currency?: CurrencyData;
    currencyOptions?: Array<CurrencyData>;
    expenseAccountOptions?: Array<GLAccountData>;
    feeFrequency?: EnumOptionData;
    feeFrequencyOptions?: Array<EnumOptionData>;
    feeInterval?: number;
    feeOnMonthDay?: ChargeFeeOnMonthDay;
    freeWithdrawal?: boolean;
    freeWithdrawalChargeFrequency?: number;
    id?: number;
    incomeOrLiabilityAccount?: GLAccountData;
    incomeOrLiabilityAccountOptions?: { [key: string]: Array<GLAccountData>; };
    isPaymentType?: boolean;
    loanChargeCalculationTypeOptions?: Array<EnumOptionData>;
    loanChargeTimeTypeOptions?: Array<EnumOptionData>;
    maxCap?: number;
    minCap?: number;
    name?: string;
    overdueInstallmentCharge?: boolean;
    paymentTypeOptions?: PaymentTypeData;
    penalty?: boolean;
    restartFrequency?: number;
    restartFrequencyEnum?: number;
    savingsChargeCalculationTypeOptions?: Array<EnumOptionData>;
    savingsChargeTimeTypeOptions?: Array<EnumOptionData>;
    shareChargeCalculationTypeOptions?: Array<EnumOptionData>;
    shareChargeTimeTypeOptions?: Array<EnumOptionData>;
    taxGroup?: TaxGroupData;
    taxGroupOptions?: Array<TaxGroupData>;
}